+++ gui/packages/desktop/src/renderer/lib/notification-controller.js
@@ -2,28 +2,30 @@
 
 import { remote } from 'electron';
 import log from 'electron-log';
+import config from '../../config';
 
 import type { TunnelStateTransition } from './daemon-rpc';
 
 export default class NotificationController {
   _activeNotification: ?Notification;
   _reconnecting = false;
+  _presentedNotifications = {};
 
-  notify(tunnelState: TunnelStateTransition) {
+  notifyTunnelState(tunnelState: TunnelStateTransition) {
     switch (tunnelState.state) {
       case 'connecting':
         if (!this._reconnecting) {
-          this._show('Connecting');
+          this._showTunnelStateNotification('Connecting');
         }
         break;
       case 'connected':
-        this._show('Secured');
+        this._showTunnelStateNotification('Secured');
         break;
       case 'disconnected':
-        this._show('Unsecured');
+        this._showTunnelStateNotification('Unsecured');
         break;
       case 'blocked':
-        this._show('Blocked all connections');
+        this._showTunnelStateNotification('Blocked all connections');
         break;
       case 'disconnecting':
         switch (tunnelState.details) {
@@ -32,7 +34,7 @@ export default class NotificationController {
             // no-op
             break;
           case 'reconnect':
-            this._show('Reconnecting');
+            this._showTunnelStateNotification('Reconnecting');
             this._reconnecting = true;
             return;
         }
@@ -44,7 +46,29 @@ export default class NotificationController {
     this._reconnecting = false;
   }
 
-  _show(message: string) {
+  notifyInconsistentVersion() {
+    this._presentNotificationOnce('inconsistent-version', () => {
+      new Notification(remote.app.getName(), {
+        body: 'Inconsistent internal version information, please restart the app',
+        silent: true,
+      });
+    });
+  }
+
+  notifyUnsupportedVersion(upgradeVersion: string) {
+    this._presentNotificationOnce('unsupported-version', () => {
+      const notification = new Notification(remote.app.getName(), {
+        body: `You are running an unsupported app version. Please upgrade to ${upgradeVersion} now to ensure your security`,
+        silent: true,
+      });
+
+      notification.addEventListener('click', () => {
+        remote.shell.openExternal(config.links.download);
+      });
+    });
+  }
+
+  _showTunnelStateNotification(message: string) {
     const lastNotification = this._activeNotification;
     const sameAsLastNotification = lastNotification && lastNotification.body === message;
 
@@ -59,6 +83,7 @@ export default class NotificationController {
     newNotification.addEventListener('show', () => {
       // If the notification is closed too soon, it might still get shown. If that happens, close()
       // should be called again so that it is closed immediately.
+      // Tracking issue: https://github.com/electron/electron/issues/12887
       if (this._activeNotification !== newNotification) {
         newNotification.close();
       }
@@ -66,6 +91,14 @@ export default class NotificationController {
 
     if (lastNotification) {
       lastNotification.close();
+    }
+  }
+
+  _presentNotificationOnce(notificationName: string, presentNotification: () => void) {
+    const presented = this._presentedNotifications;
+    if (!presented[notificationName]) {
+      presented[notificationName] = true;
+      presentNotification();
     }
   }
 }